Fixed all compiler warnings

+++ Source/Graphics/divisionsmarker.cpp
@@ -30,7 +30,7 @@ void DivisionsMarker::paintEvent(QPaintEvent *)
     p.setFont(font);
 
     for (int i = 0; i * mPixelsPerBeat * mDivision < width(); i += 4) {
-        p.drawText(i * mPixelsPerBeat * mDivision, 10, QString::number(mDivision * i / 4));
+        p.drawText(int(i * mPixelsPerBeat * mDivision), 10, QString::number(mDivision * i / 4));
     }
 }
